+++ frontend-synofone/src/pages/dashboard/products/Products.jsx
@@ -3,7 +3,7 @@ import { Table, Button } from "react-bootstrap";
 import DefaultLayout from "../../../components/dashboard/DefaultLayout";
 import Api from "../../../routes/Api";
 import { useState, useEffect } from "react";
-
+import Swal from "sweetalert2";
 
 
 const Products = () => {
@@ -29,9 +29,26 @@ const Products = () => {
     const deleteData = (id) => {
         return async () => {
             try {
-                const response = await Api.delete(`/product/${id}`);
-                console.log(response.data);
-                getProducts();
+                Swal.fire({
+                    title: 'Are you sure?',
+                    text: "You won't be able to revert this!",
+                    icon: 'warning',
+                    showCancelButton: true,
+                    confirmButtonColor: '#3085d6',
+                    cancelButtonColor: '#d33',
+                    confirmButtonText: 'Yes, delete it!'
+                }).then(async (result) => {
+                    if (result.isConfirmed) {
+                        const response = await Api.delete(`/product/${id}`);
+                        console.log(response.data);
+                        getProducts();
+                        Swal.fire(
+                            'Deleted!',
+                            'Your file has been deleted.',
+                            'success'
+                        )}}
+                )
+                
             } catch (error) {
                 console.error(error)
             }
